Fix some syntax errors and remove unnecessary parameters
