+++ cypress/support/index.d.ts
@@ -25,16 +25,17 @@ declare namespace Cypress {
       isCleanup?: boolean
     ): Chainable<Subject>;
     getSource(source: string, username?: string): Chainable<Subject>;
-   createOrganisation(
-    organisation?: string,
-    username?: string,
-    public_access?: boolean
-  ): Chainable<Subject>;
-  deleteOrganisation(
-    organisation: string,
-    username?: string,
-    isCleanup?: boolean
-  ): Chainable<Subject>;
-  getOrganisation(organisation: string, username?: string): Chainable<Subject>;
+    createOrganisation(
+      organisation?: string,
+      public_access?: boolean
+    ): Chainable<Subject>;
+    deleteOrganisation(
+      organisation: string,
+      isCleanup?: boolean
+    ): Chainable<Subject>;
+    getOrganisation(
+      organisation: string,
+      username?: string
+    ): Chainable<Subject>;
   }
 }
